l05/requirements.txt
jupyter==1.0.0
matplotlib==3.2.0
pyro-ppl==1.3.0
torch==1.2.0
scikit-learn==0.22.2.post1
seaborn==0.10.0
